Include write concern with each legacy operation
Using {w:1} and deferring the full write concern until the end can work with standalone servers and replica sets, but would be unreliable for mongos (GLE may not hit all shards previously written to). Using the full write concern will be consistent with other driver implementations, even thought it comes as a cost to performance.

Since we're allowing the write concern on each legacy operation, allow for the case of {w:0}, where we may not have a result to process.